Inline DataUnionEndpoints public methods to DataUnion, move utility methods to contracts.ts and ABIs to abi.ts
