+++ src/dataunion/DataUnion.ts
@@ -1,5 +1,25 @@
+import { getAddress } from '@ethersproject/address'
 import { BigNumber } from '@ethersproject/bignumber'
-import { DataUnionEndpoints } from '../rest/DataUnionEndpoints'
+import { arrayify, hexZeroPad } from '@ethersproject/bytes'
+import { Contract } from '@ethersproject/contracts'
+import { TransactionReceipt, TransactionResponse } from '@ethersproject/providers'
+import debug from 'debug'
+import { factoryMainnetABI, dataUnionMainnetABI, dataUnionSidechainABI } from './abi'
+import {
+    fetchDataUnionMainnetAddress,
+    fetchDataUnionSidechainAddress,
+    getDataUnionSidechainAddress,
+    getMainnetContract,
+    getMainnetContractReadOnly,
+    getSidechainContract,
+    getSidechainContractReadOnly,
+    throwIfBadAddress,
+    untilWithdrawIsComplete
+} from './contracts'
+import StreamrClient from '../StreamrClient'
+import { EthereumAddress, Todo } from '../types'
+import { until, getEndpointUrl } from '../utils'
+import authFetch from '../rest/authFetch'
 
 export interface DataUnionDeployOptions {
     owner?: string,
@@ -33,16 +53,35 @@ export interface DataUnionMemberListModificationOptions {
     confirmations?: number
 }
 
+export interface DataUnionStats {
+    activeMemberCount: Todo,
+    inactiveMemberCount: Todo,
+    joinPartAgentCount: Todo,
+    totalEarnings: Todo,
+    totalWithdrawable: Todo,
+    lifetimeMemberEarnings: Todo
+}
+
+export interface MemberStats {
+    status: Todo
+    earningsBeforeLastJoin: Todo
+    lmeAtJoin: Todo
+    totalEarnings: Todo
+    withdrawableEarnings: Todo
+}
+
+const log = debug('StreamrClient::DataUnion')
+
 export class DataUnion {
 
     contractAddress: string
     sidechainAddress: string
-    dataUnionEndpoints: DataUnionEndpoints
+    client: StreamrClient
 
-    constructor(contractAddress: string, sidechainAddress: string|undefined, dataUnionEndpoints: DataUnionEndpoints) {
+    constructor(contractAddress: string, sidechainAddress: string|undefined, client: StreamrClient) {
+        this.client = client
         this.contractAddress = contractAddress
-        this.sidechainAddress = sidechainAddress || dataUnionEndpoints.calculateDataUnionSidechainAddress(contractAddress)
-        this.dataUnionEndpoints = dataUnionEndpoints
+        this.sidechainAddress = sidechainAddress || this.calculateDataUnionSidechainAddress(contractAddress)
     }
 
     getAddress() {
@@ -55,85 +94,485 @@ export class DataUnion {
 
     // Member functions
 
-    async join(secret?: string) {
-        return this.dataUnionEndpoints.join(secret, this.contractAddress)
+    /**
+     * Send a joinRequest, or get into data union instantly with a data union secret
+     */
+    async join(secret?: string): Promise<Todo> {
+        const memberAddress = this.client.getAddress() as string
+        const body: any = {
+            memberAddress
+        }
+        if (secret) { body.secret = secret }
+
+        const url = getEndpointUrl(this.client.options.restUrl, 'dataunions', this.contractAddress, 'joinRequests')
+        const response = await authFetch(
+            url,
+            this.client.session,
+            {
+                method: 'POST',
+                body: JSON.stringify(body),
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            },
+        )
+        if (secret) {
+            await until(async () => this.isMember(memberAddress))
+        }
+        return response
     }
 
-    async isMember(memberAddress: string) {
-        return this.dataUnionEndpoints.isMember(memberAddress, this.contractAddress)
+    async isMember(memberAddress: EthereumAddress): Promise<boolean> {
+        const address = getAddress(memberAddress)
+        const duSidechain = await getSidechainContractReadOnly(this.contractAddress, this.client)
+        const ACTIVE = 1 // memberData[0] is enum ActiveStatus {None, Active, Inactive}
+        const memberData = await duSidechain.memberData(address)
+        const state = memberData[0]
+        return (state === ACTIVE)
     }
 
-    async withdrawAll(options?: DataUnionWithdrawOptions) {
-        return this.dataUnionEndpoints.withdrawAll(this.contractAddress, options)
+    /**
+     * Withdraw all your earnings
+     * @returns receipt once withdraw is complete (tokens are seen in mainnet)
+     */
+    async withdrawAll(options?: DataUnionWithdrawOptions): Promise<TransactionReceipt> {
+        const recipientAddress = this.client.getAddress()
+        const tr = await untilWithdrawIsComplete(
+            this.client,
+            () => this.getWithdrawAllTx(),
+            () => this.client.getTokenBalance(recipientAddress),
+            options
+        )
+        return tr
     }
 
-    async withdrawAllTo(recipientAddress: string, options?: DataUnionWithdrawOptions) {
-        return this.dataUnionEndpoints.withdrawAllTo(recipientAddress, options, this.contractAddress)
+    /**
+     * Get the tx promise for withdrawing all your earnings
+     * @returns await on call .wait to actually send the tx
+     */
+    private async getWithdrawAllTx(): Promise<TransactionResponse> {
+        const signer = await this.client.ethereum.getSidechainSigner()
+        const address = await signer.getAddress()
+        const duSidechain = await getSidechainContract(this.contractAddress, this.client)
+
+        const withdrawable = await duSidechain.getWithdrawableEarnings(address)
+        if (withdrawable.eq(0)) {
+            throw new Error(`${address} has nothing to withdraw in (sidechain) data union ${duSidechain.address}`)
+        }
+
+        if (this.client.options.minimumWithdrawTokenWei && withdrawable.lt(this.client.options.minimumWithdrawTokenWei)) {
+            throw new Error(`${address} has only ${withdrawable} to withdraw in `
+                + `(sidechain) data union ${duSidechain.address} (min: ${this.client.options.minimumWithdrawTokenWei})`)
+        }
+        return duSidechain.withdrawAll(address, true) // sendToMainnet=true
     }
 
-    async signWithdrawAllTo(recipientAddress: string) {
-        return this.dataUnionEndpoints.signWithdrawAllTo(recipientAddress, this.contractAddress)
+    /**
+     * Withdraw earnings and "donate" them to the given address
+     * @returns get receipt once withdraw is complete (tokens are seen in mainnet)
+     */
+    async withdrawAllTo(
+        recipientAddress: EthereumAddress,
+        options: DataUnionWithdrawOptions|undefined
+    ): Promise<TransactionReceipt> {
+        const to = getAddress(recipientAddress) // throws if bad address
+        const tr = await untilWithdrawIsComplete(
+            this.client,
+            () => this.getWithdrawAllToTx(to),
+            () => this.client.getTokenBalance(to),
+            options
+        )
+        return tr
     }
 
-    async signWithdrawAmountTo(recipientAddress: string, amountTokenWei: BigNumber|number|string) {
-        return this.dataUnionEndpoints.signWithdrawAmountTo(recipientAddress, amountTokenWei, this.contractAddress)
+    /**
+     * Withdraw earnings and "donate" them to the given address
+     * @param recipientAddress - the address to receive the tokens
+     * @returns await on call .wait to actually send the tx
+     */
+    private async getWithdrawAllToTx(recipientAddress: EthereumAddress): Promise<TransactionResponse> {
+        const signer = await this.client.ethereum.getSidechainSigner()
+        const address = await signer.getAddress()
+        const duSidechain = await getSidechainContract(this.contractAddress, this.client)
+        const withdrawable = await duSidechain.getWithdrawableEarnings(address)
+        if (withdrawable.eq(0)) {
+            throw new Error(`${address} has nothing to withdraw in (sidechain) data union ${duSidechain.address}`)
+        }
+        return duSidechain.withdrawAllTo(recipientAddress, true) // sendToMainnet=true
+    }
+
+    /**
+     * Member can sign off to "donate" all earnings to another address such that someone else
+     *   can submit the transaction (and pay for the gas)
+     * This signature is only valid until next withdrawal takes place (using this signature or otherwise).
+     * Note that while it's a "blank cheque" for withdrawing all earnings at the moment it's used, it's
+     *   invalidated by the first withdraw after signing it. In other words, any signature can be invalidated
+     *   by making a "normal" withdraw e.g. `await streamrClient.withdrawAll()`
+     * Admin can execute the withdraw using this signature: ```
+     *   await adminStreamrClient.withdrawAllToSigned(memberAddress, recipientAddress, signature)
+     * ```
+     * @param recipientAddress - the address authorized to receive the tokens
+     * @returns signature authorizing withdrawing all earnings to given recipientAddress
+     */
+    async signWithdrawAllTo(recipientAddress: EthereumAddress): Promise<string> {
+        return this.signWithdrawAmountTo(recipientAddress, BigNumber.from(0))
+    }
+
+    /**
+     * Member can sign off to "donate" specific amount of earnings to another address such that someone else
+     *   can submit the transaction (and pay for the gas)
+     * This signature is only valid until next withdrawal takes place (using this signature or otherwise).
+     * @param recipientAddress - the address authorized to receive the tokens
+     * @param amountTokenWei - that the signature is for (can't be used for less or for more)
+     * @returns signature authorizing withdrawing all earnings to given recipientAddress
+     */
+    async signWithdrawAmountTo(
+        recipientAddress: EthereumAddress,
+        amountTokenWei: BigNumber|number|string
+    ): Promise<string> {
+        const to = getAddress(recipientAddress) // throws if bad address
+        const signer = this.client.ethereum.getSigner() // it shouldn't matter if it's mainnet or sidechain signer since key should be the same
+        const address = await signer.getAddress()
+        const duSidechain = await getSidechainContractReadOnly(this.contractAddress, this.client)
+        const memberData = await duSidechain.memberData(address)
+        if (memberData[0] === '0') { throw new Error(`${address} is not a member in Data Union (sidechain address ${duSidechain.address})`) }
+        const withdrawn = memberData[3]
+        // @ts-expect-error
+        const message = to + hexZeroPad(amountTokenWei, 32).slice(2) + duSidechain.address.slice(2) + hexZeroPad(withdrawn, 32).slice(2)
+        const signature = await signer.signMessage(arrayify(message))
+        return signature
     }
 
     // Query functions
 
-    async getStats() {
-        return this.dataUnionEndpoints.getStats(this.contractAddress)
+    async getStats(): Promise<DataUnionStats> {
+        const duSidechain = await getSidechainContractReadOnly(this.contractAddress, this.client)
+        const [
+            totalEarnings,
+            totalEarningsWithdrawn,
+            activeMemberCount,
+            inactiveMemberCount,
+            lifetimeMemberEarnings,
+            joinPartAgentCount,
+        ] = await duSidechain.getStats()
+        const totalWithdrawable = totalEarnings.sub(totalEarningsWithdrawn)
+        return {
+            activeMemberCount,
+            inactiveMemberCount,
+            joinPartAgentCount,
+            totalEarnings,
+            totalWithdrawable,
+            lifetimeMemberEarnings,
+        }
     }
 
-    async getMemberStats(memberAddress: string) {
-        return this.dataUnionEndpoints.getMemberStats(memberAddress, this.contractAddress)
+    /**
+     * Get stats of a single data union member
+     */
+    async getMemberStats(memberAddress: EthereumAddress): Promise<MemberStats> {
+        const address = getAddress(memberAddress)
+        // TODO: use duSidechain.getMemberStats(address) once it's implemented, to ensure atomic read
+        //        (so that memberData is from same block as getEarnings, otherwise withdrawable will be foobar)
+        const duSidechain = await getSidechainContractReadOnly(this.contractAddress, this.client)
+        const mdata = await duSidechain.memberData(address)
+        const total = await duSidechain.getEarnings(address).catch(() => 0)
+        const withdrawnEarnings = mdata[3].toString()
+        const withdrawable = total ? total.sub(withdrawnEarnings) : 0
+        return {
+            status: ['unknown', 'active', 'inactive', 'blocked'][mdata[0]],
+            earningsBeforeLastJoin: mdata[1].toString(),
+            lmeAtJoin: mdata[2].toString(),
+            totalEarnings: total.toString(),
+            withdrawableEarnings: withdrawable.toString(),
+        }
     }
 
-    async getWithdrawableEarnings(memberAddress: string) {
-        return this.dataUnionEndpoints.getWithdrawableEarnings(memberAddress, this.contractAddress)
+    /**
+     * Get the amount of tokens the member would get from a successful withdraw
+     */
+    async getWithdrawableEarnings(memberAddress: EthereumAddress): Promise<BigNumber> {
+        const address = getAddress(memberAddress)
+        const duSidechain = await getSidechainContractReadOnly(this.contractAddress, this.client)
+        return duSidechain.getWithdrawableEarnings(address)
     }
 
-    async getAdminFee() {
-        return this.dataUnionEndpoints.getAdminFee(this.contractAddress)
+    /**
+     * Get data union admin fee fraction (between 0.0 and 1.0) that admin gets from each revenue event
+     */
+    async getAdminFee(): Promise<number> {
+        const duMainnet = getMainnetContractReadOnly(this.contractAddress, this.client)
+        const adminFeeBN = await duMainnet.adminFeeFraction()
+        return +adminFeeBN.toString() / 1e18
     }
 
-    async getAdminAddress() {
-        return this.dataUnionEndpoints.getAdminAddress(this.contractAddress)
+    async getAdminAddress(): Promise<Todo> {
+        const duMainnet = getMainnetContractReadOnly(this.contractAddress, this.client)
+        return duMainnet.owner()
     }
 
-    async getVersion() {
-        return this.dataUnionEndpoints.getVersion(this.contractAddress)
+    /**
+     * Figure out if given mainnet address is old DataUnion (v 1.0) or current 2.0
+     * NOTE: Current version of streamr-client-javascript can only handle current version!
+     */
+    async getVersion(): Promise<number> {
+        const a = getAddress(this.contractAddress) // throws if bad address
+        const provider = this.client.ethereum.getMainnetProvider()
+        const du = new Contract(a, [{
+            name: 'version',
+            inputs: [],
+            outputs: [{ type: 'uint256' }],
+            stateMutability: 'view',
+            type: 'function'
+        }], provider)
+        try {
+            const version = await du.version()
+            return +version
+        } catch (e) {
+            // "not a data union"
+            return 0
+        }
     }
 
     // Admin functions
 
-    async createSecret(name: string = 'Untitled Data Union Secret') {
-        return this.dataUnionEndpoints.createSecret(this.contractAddress, name)
+    /**
+     * Add a new data union secret
+     */
+    async createSecret(name: string = 'Untitled Data Union Secret'): Promise<string> {
+        const duAddress = getAddress(this.contractAddress) // throws if bad address
+        const url = getEndpointUrl(this.client.options.restUrl, 'dataunions', duAddress, 'secrets')
+        const res = await authFetch(
+            url,
+            this.client.session,
+            {
+                method: 'POST',
+                body: JSON.stringify({
+                    name
+                }),
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+            },
+        )
+        return res.secret
     }
 
-    async addMembers(memberAddressList: string[], options?: DataUnionMemberListModificationOptions) {
-        return this.dataUnionEndpoints.addMembers(memberAddressList, options, this.contractAddress)
+    /**
+     * Add given Ethereum addresses as data union members
+     */
+    async addMembers(
+        memberAddressList: string[],
+        options: DataUnionMemberListModificationOptions = {}
+    ): Promise<TransactionReceipt> {
+        const members = memberAddressList.map(getAddress) // throws if there are bad addresses
+        const duSidechain = await getSidechainContract(this.contractAddress, this.client)
+        const tx = await duSidechain.addMembers(members)
+        // TODO: wrap promise for better error reporting in case tx fails (parse reason, throw proper error)
+        const { confirmations = 1 } = options
+        return tx.wait(confirmations)
     }
 
-    async removeMembers(memberAddressList: string[], options?: DataUnionMemberListModificationOptions) {
-        return this.dataUnionEndpoints.removeMembers(memberAddressList, options, this.contractAddress)
+    /**
+     * Remove given members from data union
+     */
+    async removeMembers(
+        memberAddressList: string[],
+        options: DataUnionMemberListModificationOptions = {},
+    ): Promise<TransactionReceipt> {
+        const members = memberAddressList.map(getAddress) // throws if there are bad addresses
+        const duSidechain = await getSidechainContract(this.contractAddress, this.client)
+        const tx = await duSidechain.partMembers(members)
+        // TODO: wrap promise for better error reporting in case tx fails (parse reason, throw proper error)
+        const { confirmations = 1 } = options
+        return tx.wait(confirmations)
     }
 
-    async withdrawAllToMember(memberAddress: string, options?: DataUnionWithdrawOptions) {
-        return this.dataUnionEndpoints.withdrawAllToMember(memberAddress, options, this.contractAddress)
+    /**
+     * Admin: withdraw earnings (pay gas) on behalf of a member
+     * TODO: add test
+     * @param memberAddress - the other member who gets their tokens out of the Data Union
+     * @returns Receipt once withdraw transaction is confirmed
+     */
+    async withdrawAllToMember(
+        memberAddress: EthereumAddress,
+        options?: DataUnionWithdrawOptions
+    ): Promise<TransactionReceipt> {
+        const address = getAddress(memberAddress) // throws if bad address
+        const tr = await untilWithdrawIsComplete(
+            this.client,
+            () => this.getWithdrawAllToMemberTx(address),
+            () => this.client.getTokenBalance(address),
+            options
+        )
+        return tr
     }
 
-    async withdrawAllToSigned(memberAddress: string, recipientAddress: string, signature: string, options?: DataUnionWithdrawOptions) {
-        return this.dataUnionEndpoints.withdrawAllToSigned(memberAddress, recipientAddress, signature, options, this.contractAddress)
+    /**
+     * Admin: get the tx promise for withdrawing all earnings on behalf of a member
+     * @param memberAddress - the other member who gets their tokens out of the Data Union
+     * @returns await on call .wait to actually send the tx
+     */
+    private async getWithdrawAllToMemberTx(memberAddress: EthereumAddress): Promise<TransactionResponse> {
+        const a = getAddress(memberAddress) // throws if bad address
+        const duSidechain = await getSidechainContract(this.contractAddress, this.client)
+        return duSidechain.withdrawAll(a, true) // sendToMainnet=true
     }
 
-    async setAdminFee(newFeeFraction: number) {
-        return this.dataUnionEndpoints.setAdminFee(newFeeFraction, this.contractAddress)
+    /**
+     * Admin: Withdraw a member's earnings to another address, signed by the member
+     * @param memberAddress - the member whose earnings are sent out
+     * @param recipientAddress - the address to receive the tokens in mainnet
+     * @param signature - from member, produced using signWithdrawAllTo
+     * @returns receipt once withdraw transaction is confirmed
+     */
+    async withdrawAllToSigned(
+        memberAddress: EthereumAddress,
+        recipientAddress: EthereumAddress,
+        signature: string,
+        options?: DataUnionWithdrawOptions
+    ): Promise<TransactionReceipt> {
+        const from = getAddress(memberAddress) // throws if bad address
+        const to = getAddress(recipientAddress)
+        const tr = await untilWithdrawIsComplete(
+            this.client,
+            () => this.getWithdrawAllToSignedTx(from, to, signature),
+            () => this.client.getTokenBalance(to),
+            options
+        )
+        return tr
+    }
+
+    /**
+     * Admin: Withdraw a member's earnings to another address, signed by the member
+     * @param memberAddress - the member whose earnings are sent out
+     * @param recipientAddress - the address to receive the tokens in mainnet
+     * @param signature - from member, produced using signWithdrawAllTo
+     * @returns await on call .wait to actually send the tx
+     */
+    private async getWithdrawAllToSignedTx(
+        memberAddress: EthereumAddress,
+        recipientAddress: EthereumAddress,
+        signature: string,
+    ): Promise<TransactionResponse> {
+        const duSidechain = await getSidechainContract(this.contractAddress, this.client)
+        return duSidechain.withdrawAllToSigned(memberAddress, recipientAddress, true, signature) // sendToMainnet=true
+    }
+
+    /**
+     * Admin: set admin fee (between 0.0 and 1.0) for the data union
+     */
+    async setAdminFee(newFeeFraction: number): Promise<Todo> {
+        if (newFeeFraction < 0 || newFeeFraction > 1) {
+            throw new Error('newFeeFraction argument must be a number between 0...1, got: ' + newFeeFraction)
+        }
+        const adminFeeBN = BigNumber.from((newFeeFraction * 1e18).toFixed()) // last 2...3 decimals are going to be gibberish
+        const duMainnet = getMainnetContract(this.contractAddress, this.client)
+        const tx = await duMainnet.setAdminFee(adminFeeBN)
+        return tx.wait()
+    }
+
+    /**
+     * Create a new DataUnionMainnet contract to mainnet with DataUnionFactoryMainnet
+     * This triggers DataUnionSidechain contract creation in sidechain, over the bridge (AMB)
+     * @return that resolves when the new DU is deployed over the bridge to side-chain
+     */
+    static async _deployContract(options: DataUnionDeployOptions = {}, client: StreamrClient): Promise<Contract> {
+        const {
+            owner,
+            joinPartAgents,
+            dataUnionName,
+            adminFee = 0,
+            sidechainPollingIntervalMs = 1000,
+            sidechainRetryTimeoutMs = 600000,
+            confirmations = 1,
+            gasPrice
+        } = options
+
+        let duName = dataUnionName
+        if (!duName) {
+            duName = `DataUnion-${Date.now()}` // TODO: use uuid
+            log(`dataUnionName generated: ${duName}`)
+        }
+
+        if (adminFee < 0 || adminFee > 1) { throw new Error('options.adminFeeFraction must be a number between 0...1, got: ' + adminFee) }
+        const adminFeeBN = BigNumber.from((adminFee * 1e18).toFixed()) // last 2...3 decimals are going to be gibberish
+
+        const mainnetProvider = client.ethereum.getMainnetProvider()
+        const mainnetWallet = client.ethereum.getSigner()
+        const sidechainProvider = client.ethereum.getSidechainProvider()
+
+        const ownerAddress = (owner) ? getAddress(owner) : client.getAddress()
+
+        let agentAddressList
+        if (Array.isArray(joinPartAgents)) {
+            // getAddress throws if there's an invalid address in the array
+            agentAddressList = joinPartAgents.map(getAddress)
+        } else {
+            // streamrNode needs to be joinPartAgent so that EE join with secret works (and join approvals from Marketplace UI)
+            agentAddressList = [ownerAddress]
+            if (client.options.streamrNodeAddress) {
+                agentAddressList.push(getAddress(client.options.streamrNodeAddress))
+            }
+        }
+
+        const deployerAddress = client.getAddress()
+        const duMainnetAddress = await fetchDataUnionMainnetAddress(client, duName, deployerAddress)
+        const duSidechainAddress = await fetchDataUnionSidechainAddress(client, duMainnetAddress)
+
+        if (await mainnetProvider.getCode(duMainnetAddress) !== '0x') {
+            throw new Error(`Mainnet data union "${duName}" contract ${duMainnetAddress} already exists!`)
+        }
+
+        const factoryMainnetAddress = throwIfBadAddress(
+            client.options.factoryMainnetAddress!,
+            'StreamrClient.options.factoryMainnetAddress'
+        )
+        if (await mainnetProvider.getCode(factoryMainnetAddress) === '0x') {
+            throw new Error(`Data union factory contract not found at ${factoryMainnetAddress}, check StreamrClient.options.factoryMainnetAddress!`)
+        }
+
+        // function deployNewDataUnion(address owner, uint256 adminFeeFraction, address[] agents, string duName)
+        const factoryMainnet = new Contract(factoryMainnetAddress!, factoryMainnetABI, mainnetWallet)
+        const ethersOptions: any = {}
+        if (gasPrice) {
+            ethersOptions.gasPrice = gasPrice
+        }
+        const tx = await factoryMainnet.deployNewDataUnion(
+            ownerAddress,
+            adminFeeBN,
+            agentAddressList,
+            duName,
+            ethersOptions
+        )
+        const tr = await tx.wait(confirmations)
+
+        log(`Data Union "${duName}" (mainnet: ${duMainnetAddress}, sidechain: ${duSidechainAddress}) deployed to mainnet, waiting for side-chain...`)
+        await until(
+            async () => await sidechainProvider.getCode(duSidechainAddress) !== '0x',
+            sidechainRetryTimeoutMs,
+            sidechainPollingIntervalMs
+        )
+
+        const dataUnion = new Contract(duMainnetAddress, dataUnionMainnetABI, mainnetWallet)
+        // @ts-expect-error
+        dataUnion.deployTxReceipt = tr
+        // @ts-expect-error
+        dataUnion.sidechain = new Contract(duSidechainAddress, dataUnionSidechainABI, sidechainProvider)
+        return dataUnion
     }
 
     // Internal functions
 
     async _getContract() {
-        return this.dataUnionEndpoints.getContract(this.contractAddress)
+        const ret = getMainnetContract(this.contractAddress, this.client)
+        // @ts-expect-error
+        ret.sidechain = await getSidechainContract(this.contractAddress, this.client)
+        return ret
+    }
+
+    // TODO inline this function?
+    private calculateDataUnionSidechainAddress(duMainnetAddress: EthereumAddress) {
+        const address = getAddress(duMainnetAddress) // throws if bad address
+        return getDataUnionSidechainAddress(this.client, address)
     }
 }
